third_party/WebKit/LayoutTests/http/tests/credentialmanager/localcredential-basics.html

<!DOCTYPE html>
<title>Credential Manager: LocalCredential basics.</title>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="/serviceworker/resources/interfaces.js"></script>
<script>
test(function() {
    var credential = new LocalCredential('id', 'pencil', 'name', 'https://example.com/avatar.png');

    verify_interface('LocalCredential', credential, {
        formData: 'object',
        id: 'string',
        name: 'string',
        avatarURL: 'string',
        password: 'string',
        type: 'string'
    });

    assert_true(credential.formData instanceof FormData);
    assert_equals(credential.id, 'id');
    assert_equals(credential.name, 'name');
    assert_equals(credential.avatarURL, 'https://example.com/avatar.png');
    assert_equals(credential.password, 'pencil');
    assert_equals(credential.type, 'password');
}, 'Interfaces and attributes of LocalCredential');

test(function() {
    assert_throws(new SyntaxError(), function () {
        var credential = new LocalCredential('id', 'pencil', 'name', '-');
    });
}, 'Construct a LocalCredential with an invalid avatar URL.');

test(function() {
    var credential = new LocalCredential('id', 'pencil', 'name');

    assert_equals(credential.id, 'id');
    assert_equals(credential.name, 'name');
    assert_equals(credential.avatarURL, '');
    assert_equals(credential.password, 'pencil');
    assert_equals(credential.type, 'password');

}, 'Construct a LocalCredential with an empty avatar URL.');

test(function() {
    var credential = new LocalCredential('id', 'pencil');

    assert_equals(credential.id, 'id');
    assert_equals(credential.name, '');
    assert_equals(credential.avatarURL, '');
    assert_equals(credential.password, 'pencil');
    assert_equals(credential.type, 'password');

}, 'Construct a LocalCredential with an empty name and avatar URL.');
</script>
